fix: long 64 位常量支持 — Token.ival/AstNode.ival 改为 long long
016_long 全部通过。关键变更：
- Token.ival: int → long long
- AstNode.ival: int → long long
- 新增 mov_rax_imm64() 发射 64 位立即数
- AST_CONSTANT 根据值范围选择 mov_eax_imm(32-bit) 或 mov_rax_imm64(64-bit)
- eval_const_expr 返回类型 int → long long

至此所有 93 个编译器测试全部通过。

+++ app/compiler/parse.c
@@ -1064,7 +1064,7 @@ static int parse_struct_type(Parser *p, StructType *out) {
 /* ─── 类型说明符 ─── */
 
 /* evaluate compile-time constant expression */
-static int eval_const_expr(AstNode *n) {
+static long long eval_const_expr(AstNode *n) {
     if (!n) return 0;
     switch (n->kind) {
     case AST_CONSTANT: return n->ival;
